requirements.txt
# Core
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Deep Learning
tensorflow==2.13.0

# Chemistry
rdkit-pypi==2023.9.5

# Visualization
matplotlib==3.7.2